Add files via upload
The first file 1_1 : We extract all data from the parquet files and we save it following a certain structure : we split at first by stock ticker then by trade and bbo , then, within each directory, we group csv files in directories by month. We also do some initial inspection of what a random csv file contains, it turns out that it has the same structure of the TRTH files 

In the second file, we do some parallel computation to read, merge and save the data files located in the mentioned directories, the file is still in progress, and need to be tested, ideally, what I want to do, is to permit to extract data in a flexible way: (of course in a fast way): can select a range of dates to read the files both for trades and bbo, offer a possible merge between trade and bbo, return a data either in cleaned version or in raw version: This would allow us to rapidly manipulate the data easily in the time range we want to focus on: That's it for today.