Add files via upload
The functions work well, 2 problems spotted : 
One related to parquet or pyarrow : after saving the file , I don't know why when I want to read the data it doesn't work  (I tried using different methods one with pyarrow and one with fast.parquet both gave different type of error )
I tried to change the python version or the packages versions and yet nothing . 
The second one also is related to vaex which I didn't use, because of a problem in packages. 
I will try tomorrow to check what is the solutions for such problem : Once fixed, we will easily access the data and make the computations we would like to do.